refactor(serialization): clean up and move into its own class
