+++ src/main/java/com/conveyal/r5/streets/LinkedPointSet.java
@@ -395,6 +395,7 @@ public class LinkedPointSet implements Serializable {
      */
     public void makeStopToPointDistanceTables(Geometry treeRebuildZone) {
         LOG.info("Creating distance tables from each transit stop to PointSet points.");
+        // FIXME this is wasting a lot of memory and not needed for gridded pointsets - overload for gridded and freeform PointSets
         pointSet.createSpatialIndexAsNeeded();
         if (treeRebuildZone != null) {
             LOG.info("Selectively computing tables for only those stops that might be affected by the scenario.");
